FilterLineReachabilityRows: move into filterlinereachability package
Close batfish/batfish#2318
