+++ projects/question/src/test/java/org/batfish/question/filterlinereachability/FilterLineReachabilityTest.java
@@ -3,18 +3,18 @@ package org.batfish.question.filterlinereachability;
 import static org.batfish.datamodel.IpAccessListLine.acceptingHeaderSpace;
 import static org.batfish.datamodel.IpAccessListLine.rejectingHeaderSpace;
 import static org.batfish.datamodel.LineAction.PERMIT;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.COLUMN_METADATA;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.COL_ADDITIONAL_INFO;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.COL_BLOCKING_LINES;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.COL_DIFF_ACTION;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.COL_REASON;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.COL_SOURCES;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.COL_UNREACHABLE_LINE;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.COL_UNREACHABLE_LINE_ACTION;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.Reason.BLOCKING_LINES;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.Reason.CYCLICAL_REFERENCE;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.Reason.INDEPENDENTLY_UNMATCHABLE;
-import static org.batfish.datamodel.answers.FilterLineReachabilityRows.Reason.UNDEFINED_REFERENCE;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.COLUMN_METADATA;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.COL_ADDITIONAL_INFO;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.COL_BLOCKING_LINES;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.COL_DIFF_ACTION;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.COL_REASON;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.COL_SOURCES;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.COL_UNREACHABLE_LINE;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.COL_UNREACHABLE_LINE_ACTION;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.Reason.BLOCKING_LINES;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.Reason.CYCLICAL_REFERENCE;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.Reason.INDEPENDENTLY_UNMATCHABLE;
+import static org.batfish.question.filterlinereachability.FilterLineReachabilityRows.Reason.UNDEFINED_REFERENCE;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
